+++ CHANGES.md
@@ -1,3 +1,7 @@
+## Unreleased
+
+- Removed `maailma.encrypt` namespace
+
 ## 0.2.0 (2015-11-07)
 
 Initial public release.